refactor(cache): split template source cache policy
