+++ .sampo/changesets/789-template-source-cache-policy.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split external template cache policy helpers into a focused runtime module.
